python-services/requirements.txt
flask==3.1.0
flask-cors==5.0.0
torch==2.6.0
torchvision==0.21.0
opencv-python==4.11.0.86
Pillow==11.1.0
matplotlib==3.10.0
numpy==2.2.3